Disable AOT by Default and Fix Failing Tests
This pull request addresses the issue of disabling AOT (Ahead-of-Time) compilation by default and fixing as many failing tests as possible. The AOT flag has been set to false by default, and the code has been updated to ensure compatibility with this setting. Additionally, timeouts of 5 seconds have been added to tests that were hanging indefinitely to prevent them from stalling the test suite. This should improve the stability and reliability of the test runs.